Formalize worklog by using proper go types and tests (#5)
Formalize worklog by using proper go types and tests

The entire thing was redone because of a bug where sometimes the
computed next day for `-friday` was incorrect. Rather than hacking
around the bug, I instead decided to make this much more robust and add
some test coverage. There's still a lot of coverage to add, but it is
not so important now.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>

---------

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>
Co-authored-by: Lisa Seelye <lisa@users.noreply.github.com>